cv_model/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pandas
scikit-learn
joblib
openai
python-dotenv
PyPDF2
pytesseract
pdf2image
sentence-transformers
numpy
